day3/mcp/requirements.txt
mcp[cli]>=1.2.0
httpx
python-dotenv
streamlit
google-genai
pandas
plotly
